requirements.txt
DateTime==4.1.1
github3.py==0.9.6
pytz==2016.7
slacker==0.9.25
APScheduler==3.0.0

